+++ packages/viewer/src/components/scene-tree/lib/jwt.ts
@@ -1,0 +1,9 @@
+import jwtDecode, { JwtPayload } from 'jwt-decode';
+
+export interface SceneTreeJwtToken extends JwtPayload {
+  view: string;
+}
+
+export function decodeSceneTreeJwt(jwt: string): SceneTreeJwtToken {
+  return jwtDecode<SceneTreeJwtToken>(jwt);
+}
