+++ packages/viewer/tsconfig.json
@@ -9,6 +9,6 @@
     "jsxFactory": "h",
     "target": "es2017"
   },
-  "include": ["src", "types/*.d.ts"],
+  "include": ["src", "types/*.d.ts", "src/components.d.ts"],
   "exclude": ["dist", "node_modules"]
 }
